# Improve performance when ingesting