--- a/common/src/db/mod.rs
+++ b/common/src/db/mod.rs
@@ -1,3 +1,4 @@
+pub mod chunk;
 pub mod limiter;
 pub mod test;