--- a/migration/src/m0000140_create_package_version_range.rs
+++ b/migration/src/m0000140_create_package_version_range.rs
@@ -29,7 +29,7 @@ impl MigrationTrait for Migration {
                     )
                     .col(
                         ColumnDef::new(PackageVersionRange::PackageId)
-                            .integer()
+                            .uuid()
                             .not_null(),
                     )
                     .foreign_key(