--- a/migration/src/m0000160_create_fixed_package_version.rs
+++ b/migration/src/m0000160_create_fixed_package_version.rs
@@ -44,7 +44,7 @@ impl MigrationTrait for Migration {
                     )
                     .col(
                         ColumnDef::new(FixedPackageVersion::PackageVersionId)
-                            .integer()
+                            .uuid()
                             .not_null(),
                     )
                     .foreign_key(