--- a/migration/src/m0000170_create_not_affected_package_version.rs
+++ b/migration/src/m0000170_create_not_affected_package_version.rs
@@ -44,7 +44,7 @@ impl MigrationTrait for Migration {
                     )
                     .col(
                         ColumnDef::new(NotAffectedPackageVersion::PackageVersionId)
-                            .integer()
+                            .uuid()
                             .not_null(),
                     )
                     .foreign_key(